# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Serwish Api",
        "description": "Serwish description",
        "contact": {
            "email": "dev@serwish.ge"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://serwish.ge",
            "description": "serwish web-site"
        },
        {
            "url": "https://api.serwish.ge/api",
            "description": "serwish api"
        }
    ],
    "paths": {
...
        "/api/services/orders/start-order": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "start first order ",
                "operationId": "postStartORder",
                "parameters": [
                    {
                        "name": "filterable",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/StartOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "fail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiRequestException"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Services"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/services/orders/get-orders/list": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get orders ",
                "operationId": "getOrdersByUser",
                "parameters": [
                    {
                        "name": "filterable",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/GetOrdersRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Services"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/services/orders/get-order/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get orders ",
                "operationId": "getSingleOrderById",
                "parameters": [
                    {
                        "name": "filterable",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/GetOrdersRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Services"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/services/orders/create-payment": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "create payment for group ",
                "operationId": "createPaymentRequest",
                "parameters": [
                    {
                        "name": "filterable",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/CreatePaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "fail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiRequestException"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Services"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/services/orders/get-payments/list": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get payment list ",
                "operationId": "getPayments",
                "parameters": [
                    {
                        "name": "filterable",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/GetGroupPaymentsRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Services"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/services/orders/request-withdraw": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "create payment for group ",
                "operationId": "requestWithdrawal",
                "parameters": [
                    {
                        "name": "filterable",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/CreateWithdrawalRequest"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "fail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiRequestException"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserWithdrawalRequests"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/services/orders/order/message/attachment": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "create orders for group ",
                "operationId": "uploadMessageAttachment",
                "parameters": [
                    {
                        "name": "filterable",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/MediaStoreRequest"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "fail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiRequestException"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OrderGroupMessage"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApiRequestException": {
                "properties": {
                    "message": {
                        "description": "The err message",
                        "type": "string",
                        "example": "Not Found"
                    }
                },
                "type": "object"
            },
...
            "CreatePaymentRequest": {
                "properties": {
                    "amount": {
                        "description": "30.15",
                        "format": "string",
                        "default": "10.20"
                    },
                    "group_id": {
                        "description": "",
                        "format": "integer",
                        "default": "10"
                    }
                },
                "type": "object"
            },
...
            "CreateWithdrawalRequest": {
                "properties": {
                    "amount": {
                        "description": "30.15",
                        "format": "string",
                        "default": "10.20"
                    }
                },
                "type": "object"
            },
...
            "GetGroupPaymentsRequest": {
                "properties": {
                    "type": {
                        "description": "started,half_payment_requested,half_payment_approved,payment_finished,order_closed",
                        "format": "string",
                        "default": "all"
                    },
                    "group_id": {
                        "description": "",
                        "format": "integer",
                        "default": "10"
                    },
                    "perPage": {
                        "description": "",
                        "format": "integer",
                        "default": "10"
                    }
                },
                "type": "object"
            },
            "GetOrdersRequest": {
                "properties": {
                    "type": {
                        "description": "started,half_payment_requested,half_payment_approved,payment_finished,order_closed",
                        "format": "string",
                        "default": "all"
                    },
                    "perPage": {
                        "description": "",
                        "format": "integer",
                        "default": "10"
                    }
                },
                "type": "object"
            },
...
            "StartOrderRequest": {
                "properties": {
                    "service_id": {
                        "description": "service_id",
                        "format": "integer",
                        "default": 1
                    }
                },
                "type": "object"
            },
...
            "MediaStoreRequest": {
                "properties": {
                    "file": {
                        "type": "string",
                        "format": "binary"
                    },
                    "group_id": {
                        "type": "number"
                    },
                    "is_main": {
                        "type": "boolean",
                        "default": false
                    }
                },
                "type": "object"
            },
...
            "OrderGroupMessage": {},
...
            "Services": {
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "format": "int64"
                    },
                    "user_id": {
                        "title": "user_id",
                        "description": "user_id",
                        "format": "int64"
                    },
                    "price": {
                        "title": "price",
                        "description": "price",
                        "format": "double"
                    },
                    "is_active": {
                        "title": "is_active",
                        "description": "is_active",
                        "format": "boolean"
                    },
                    "contact_number": {
                        "title": "contact_number",
                        "description": "contact_number",
                        "format": "string"
                    },
                    "price_type": {
                        "title": "price_type",
                        "description": "price_type",
                        "format": "string"
                    },
                    "has_online_payment": {
                        "title": "has_online_payment",
                        "description": "has_online_payment",
                        "format": "boolean"
                    },
                    "has_serwish_quality": {
                        "title": "has_serwish_quality",
                        "description": "has_serwish_quality",
                        "format": "boolean"
                    },
                    "translated": {
                        "$ref": "#/components/schemas/ServicesTranslation"
                    }
                },
                "type": "object"
            },
...
            "UserWithdrawalRequests": {}
        },
        "securitySchemes": {
            "Password Based": {
                "type": "oauth2",
                "description": "Use a global Authorization: Bearer token",
                "in": "header",
                "flows": {
                    "password": {
                        "authorizationUrl": "/oauth/token",
                        "tokenUrl": "/oauth/token",
                        "refreshUrl": "/oauth/token",
                        "scopes": {}
                    }
                },
                "scheme": "https"
            }
        }
    },
    "security": [
        []
    ]
}